Cleaned up the vectornav and windmaster packet-enqueing  implementations to be more readable.
Added logfile rotation and periodic fsync in the recorder service. New logfile created every 20MB instead of appending to a single massive binary logfile.

Ripped out DMA TX for the vectornav. Completely unnecessary and complicates the code. Both the Windmaster and Vectornav use a standard poll-TX-flush-RX sequence to send commands to the sensors.